Added same asset Merging Logic to AssetManager, refactored UI to be based on AssetManager
