+++ src/main/java/Presenters/PanelFactory.java
@@ -1,4 +1,4 @@
-package Interfaces.GraphicsPresenter;
+package Presenters;
 
 import Entities.Assets.DataAccessInterface;
 import Entities.Containers.Portfolio;
@@ -8,13 +8,11 @@ import javax.swing.*;
 public class PanelFactory {
 
     //____________________ Variables ___________________________________________________________________________________
-    Portfolio portfolio;
     DataAccessInterface api;
 
 
     //____________________ Constructors ________________________________________________________________________________
-    public PanelFactory(Portfolio portfolio, DataAccessInterface api){
-        this.portfolio = portfolio;
+    public PanelFactory(DataAccessInterface api){
         this.api = api;
     };
 
@@ -23,19 +21,19 @@ public class PanelFactory {
     public JPanel makePanel(String type, int x, int y, int width, int height) {
 
         if (type == "Text") {
-            TextPanel text = new TextPanel(portfolio, api);
+            TextPanel text = new TextPanel(api);
             return text.getPanel(x, y, width, height);
         } else if (type == "Portfolio Value Chart") {
             PortfolioValueChartPanel growthChart = new PortfolioValueChartPanel();
             return growthChart.getPanel(x, y, width, height);
         } else if (type == "Portfolio Composition Chart") {
-            PortfolioPieChartPanel pieChart = new PortfolioPieChartPanel(portfolio, api);
+            PortfolioPieChartPanel pieChart = new PortfolioPieChartPanel(api);
             return pieChart.getPanel(x, y, width, height);
         } else if (type == "Asset Growth Chart") {
-            AssetGrowthChartPanel barChart = new AssetGrowthChartPanel(portfolio, api);
+            AssetGrowthChartPanel barChart = new AssetGrowthChartPanel(api);
             return barChart.getPanel(x, y, width, height);
         } else if (type == "Portfolio Growth Chart") {
-            PortfolioGrowthChartPanel barChart = new PortfolioGrowthChartPanel(portfolio, api);
+            PortfolioGrowthChartPanel barChart = new PortfolioGrowthChartPanel(api);
             return barChart.getPanel(x, y, width, height);
         } else if (type == "User Leaderboard") {
             UserLeaderboardChartPanel barChart = new UserLeaderboardChartPanel(api);
